feat: successfully ran IBM logo program

+++ src/core/interface.rs
@@ -25,7 +25,7 @@ pub enum Key {
 
 pub trait Display {
     fn clear(&mut self);
-    fn draw_sprite(&mut self, x: u8, y: u8, sprite: &[u8]) -> bool;
+    fn draw_sprite(&mut self, x: usize, y: usize, sprite: &[u8]) -> bool;
 }
 
 pub trait Input {
